Drive the page title being a link to the "home" page
This restructures the store title specs into a new sub-describe, and adds a spec requiring that the store title link to the "home page" of the app.

We also refactored the page objects so that there's now a `pageTitle` object which we can click on, and get the `titleText` from.

Co-authored-by: KK Lamberty <lamberty@morris.umn.edu>