Get unit tests to pass without warnings
The unit tests were failing because the default tests expected the main page to contain `phone-store app is running!`, which is no longer true. There was also a warning because the top-bar spec didn't have the `RouterTestingModule` in its imports. 

This fixes those issues, and documents the work.

We need to set up GitHub Actions tests so that the E2E and unit tests are being checked on every commit. If we'd done that at the start this would have shown up a lot earlier and been addressed in a more timely fashion.