BDD drive requirement that product names be links with titles
The tutorial has every product name be a link (eventually to a product-specific page), and each of these links should have a title attribute that is the name of the phone followed by `' details'`.

This commit adds tests that drive that change, and then the code from the tutorial that makes that pass.