Run the upgrade script
These are the changes generated by:

```bash
ng update @angular/core @angular/cli
```

I ran all the tests and they pass, and the app seems to load/run like it should. There's *very* little functionality at this point, though, so I'm not sure if we'd notice if something was broken.