Get the tests to pass
I was anxious about the "by hand" changes I made in adding the strictness checks so I really wanted the tests to pass, so I essentially repeated some of the work done in the `bdd-product-list` branch. Not optimal (I probably should have just cherry-picked the relevant commits), but it got the tests to pass, so I'm pretty comfortable merging this in.