Fix `getLink` so it works with strict setting
When we added the Angular 10 `--strict` settings `getLink` and calls using `.each()` broke because (for reasons we are totally unclear on) `.each()` can pass `undefined` as an argument to the callback function,
and `getLink` was typed so that `undefined` was not allowed.

After a lot of flailing we fixed this by allowed `undefined` as an input type for `getLink`, and throwing an exception if `undefined` is in fact passed in.

This seemed to minimize the amount of other stuff needed to get the code to compile and run successfully.

An alternative would be to check in lots of places for `undefined` like [in this blog post](https://www.toolsqa.com/protractor/elementarrayfinder-in-protractor/), but that seemed uglier.

Co-authored-by: K.K. Lamberty <lamberty@morris.umn.edu>