Include test for title links having details and adjust formatting
This commit includes some exploration of how to get at attributes of a tag. It also includes some updates to match the general structure of 'describe' and 'it' like we had been using in Protractor. It is not completely updated, but I'm breaking for now and wanted to push these changes.