Add `cypress-commands`
This gives us access to `.attribute()`, which is handy for our tests that wanted the `title` from the link tags. The other parts of getting that working are in previous commits, but this part I missed initially. It also adds `node`, which I think was important when I started switching to page objects.