Document E2E tests from Angular repo
The Angular repo has the starter code and the "finished" code for this tutorial, along with simple E2E tests for both of those versions. This starts the process of converting the tutorial into a checklist of tests by
documenting the tests in the Angular repository.

These are pretty limited, and actually highlight some weaknesses to the approach taken in the tutorial, so I think it was worth going through those.