Try other ways to match each product entry
There is a frustrating lack of a tidy "forEach" that lets you check each element of a list to be sure it passes the test. This new version I wrote takes into account that we can use `wrap` to let us continue with a thing in a more "cypress-y" way. That's a neat thing to know about, but it still isn't super readable. Maybe we will want to add a custom Cypress command for this (and maybe share it with the world if it's good).