Enable leveraging baseUrl
It looks like `cy.visit()` will prepend the base url for relative urls. We don't need that yet. I could not get this to be helpful for just plan visiting the baseUrl for some reason, so that may be something I just don't quite know about yet. We don't visit anything yet other than the base url, so... meh.