Add two more tests using `get`s from the Cypress playground
@KKLamberty and I played with the Cypress Playground some and found that `cy.get('.product-entry h3 a[title]')`allows us to directly access the title entries on the product entry links. This adds two tests based on that.

The one problem that needed to be addressed when turned that into a test is the fact that one _could_ have product entries that simply don't have the rest of the pattern (`h3 a`) and the test would pass because it would just ignore those entries.

The first test deals with this by confirming that there are three instances of the desired pattern (so every product entry has the pattern), and then checking that each title has the desired pattern.

The second test avoids the complication of two checks by first extracting all the product entries, and then checking that each of them matches the sub-pattern `h3 > a[title]`, etc., etc. This is nice because the `productEntry.find('h3 > a[title]')` call returns `undefined` if the given product entry doesn't have the expected structure, which then causes the test to fail. It also avoids hard-coding `3` into the test, so it shouldn't be as fragile in the face of adding to or removing from the list of products.