Get `ng e2e` to run Cypress tests
We copied various things from [@floogulinc's work on the 3601 iteration template](https://github.com/UMM-CSci-3601/3601-iteration-template/pull/119) and made some other changes based on Internet searching and it works! (Some/most/all of this is still magic to us.)

It looks like most of @floogulinc's work was based on [this Medium article](https://medium.com/@joaqcid/5-steps-to-set-cypress-as-your-e2e-test-runner-in-angular-bbaf63619a1b).

Co-authored-by: K.K. Lamberty <lamberty@morris.umn.edu>