BDD the alert action for the share buttons
This provides tests that drive the action for the share buttons. We had to use `cy.stub` to capture the calls to `window:alert` and then check that they had the correct text. All a little complex, but probably unavoidable when using alerts, which are really outside what's happening "in the app".

We then added the `(click)="share()"`to the HTML and the `share()` method to the TS from the tutorial.

To limit this to the "Share" buttons (in anticipation of having "Notify Me!" buttons as well), we ended up using `find('button:contains("Share")')`, which doesn't appear to be documented anywhere we can find. KK found a GitHub issue that somewhat incidentally described this approach, which worked really nicely: https://github.com/cypress-io/cypress/issues/3745#issuecomment-474211567

This completes part 5 of the "Template syntax" section of the first page of the tutorial, which then completes the "Template syntax" section!

Co-authored-by: K.K. Lamberty <lamberty@morris.umn.edu>